+++ data/definitions/animations.yml
@@ -1199,3 +1199,4 @@ miasmic_rush_cast: 10513
 miasmic_burst_cast: 10516
 miasmic_blitz_cast: 10524
 miasmic_barrage_cast: 10518
+charge: 811
