--- game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/magic/Charge.kts
@@ -1,22 +1,0 @@
-package world.gregs.voidps.world.interact.entity.player.combat.magic
-
-import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.hasEffect
-import world.gregs.voidps.engine.entity.item.EquipSlot
-import world.gregs.voidps.engine.entity.item.equipped
-import world.gregs.voidps.engine.event.Priority
-import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.interact.entity.combat.HitDamageModifier
-import world.gregs.voidps.world.interact.entity.combat.spell
-
-fun isMatchingArenaSpell(spell: String, cape: String): Boolean = isSaradomin(spell, cape) || isGuthix(spell, cape) || isZamorak(spell, cape)
-fun isSaradomin(spell: String, cape: String): Boolean = spell == "saradomin_strike" && cape == "saradomin_cape"
-fun isGuthix(spell: String, cape: String): Boolean = spell == "claws_of_guthix" && cape == "guthix_cape"
-fun isZamorak(spell: String, cape: String): Boolean = spell == "flames_of_zamorak" && cape == "zamorak_cape"
-
-on<HitDamageModifier>(
-    { player -> type == "spell" && player.hasEffect("charge") && isMatchingArenaSpell(player.spell, player.equipped(EquipSlot.Cape).name) },
-    priority = Priority.HIGHEST
-) { _: Player ->
-    damage += 100.0
-}
